Small fixes to plugins
- fix Targets so that if SkyCam image is changed it replots
- Misc cleanups to PolarSky and change it so that it's times are
  updated correctly when the SiteSelector changes sites or time

+++ spot/plugins/PolarSky.py
@@ -12,18 +12,14 @@ naojsoft packages
 """
 # 3rd party
 import numpy as np
-from datetime import datetime
-from dateutil import tz
 
 # ginga
-from ginga.gw import Widgets, GwHelp
+from ginga.gw import Widgets
 from ginga import GingaPlugin
+from ginga.misc import Bunch
 
 # qplan
 from qplan.util import calcpos
-
-# local
-from spot.util.polar import subaru_normalize_az
 
 
 class PolarSky(GingaPlugin.LocalPlugin):
@@ -32,11 +28,11 @@ class PolarSky(GingaPlugin.LocalPlugin):
         # superclass defines some variables for us, like logger
         super().__init__(fv, fitsimage)
 
-        # get PloarSky preferences
+        # get PolarSky preferences
         prefs = self.fv.get_preferences()
         self.settings = prefs.create_category('plugin_PolarSky')
         self.settings.add_defaults(image_radius=1850,
-                                   table_timer_update_interval=60.0)
+                                   times_update_interval=1.0)
         self.settings.load(onError='silent')
 
         self.base_circ = None
@@ -50,67 +46,77 @@ class PolarSky(GingaPlugin.LocalPlugin):
         self.orig_bg = self.viewer.get_bg()
         self.orig_fg = self.viewer.get_fg()
 
-        self.tmr_table = GwHelp.Timer(duration=self.settings
-                                      ['table_timer_update_interval'])
-        self.tmr_table.add_callback('expired', self.update_table_timer_cb)
-
-        self.tmr_sunmoon = GwHelp.Timer(duration=(60*60*12))  # 12 hours total
-        self.tmr_sunmoon.add_callback('expired', self.update_sunmoon_timer_cb)
+        # these are set in callbacks
+        self.site_obj = None
+        self.dt_utc = None
+        self.cur_tz = None
 
         self.gui_up = False
 
     def get_time_info(self):
-        self.dt_utc = datetime.utcnow().replace(tzinfo=tz.UTC)
-        self.dt = self.dt_utc.astimezone(self.tz)
+        dt = self.dt_utc.astimezone(self.cur_tz)
+        info = Bunch.Bunch()
 
-        self.date_current = self.dt.strftime("%Y-%m-%d")
-        self.local_current = self.dt.strftime("%H:%M")
-        self.utc = self.dt_utc.strftime("%H:%M")
-
-        # STILL NEED LST
+        # TODO: still need LST
+        tzname = self.cur_tz.tzname(dt)
+        info.update(dict(date_current = dt.strftime("%Y-%m-%d"),
+                         local_current = dt.strftime("%H:%M:%S") + f" [{tzname}]",
+                         utc = self.dt_utc.strftime("%H:%M:%S")))
+        return info
 
     def get_sunmoon_info(self):
-        # Sun rise/set info
-        self.sun_set = (self.site.sunset(self.dt)).strftime("%H:%M:%S")
-        self.nautical_set = (self.site.evening_twilight_12(
-                             self.dt)).strftime("%H:%M:%S")
-        self.astronomical_set = (self.site.evening_twilight_18(
-                                 self.dt)).strftime("%H:%M:%S")
-        self.sun_rise = (self.site.sunrise(self.dt)).strftime("%H:%M:%S")
-        self.nautical_rise = (self.site.morning_twilight_12(
-                              self.dt)).strftime("%H:%M:%S")
-        self.astronomical_rise = (self.site.morning_twilight_18(
-                                  self.dt)).strftime("%H:%M:%S")
+        dt = self.dt_utc.astimezone(self.cur_tz)
+        site = self.site_obj.observer
+        info = Bunch.Bunch()
 
-        # Moon info here
-        moon_data = calcpos.Moon.calc(self.site, self.dt)
-        self.moon_rise = (self.site.moon_rise(self.dt)).strftime("%H:%M:%S")
-        self.moon_set = (self.site.moon_set(self.dt)).strftime("%H:%M:%S")
-        self.moon_illum = str("%.2f" % ((self.site.moon_phase(
-                              self.dt))*100))+"%"
-        self.moon_RA = moon_data.ra
-        self.moon_DEC = moon_data.dec
+        info.update(dict(
+            # Sun rise/set info
+            sun_set = (site.sunset(dt)).strftime("%H:%M:%S"),
+            nautical_set = (site.evening_twilight_12(
+                dt)).strftime("%H:%M:%S"),
+            astronomical_set = (site.evening_twilight_18(
+                dt)).strftime("%H:%M:%S"),
+            sun_rise = (site.sunrise(dt)).strftime("%H:%M:%S"),
+            nautical_rise = (site.morning_twilight_12(
+                dt)).strftime("%H:%M:%S"),
+            astronomical_rise = (site.morning_twilight_18(
+                dt)).strftime("%H:%M:%S")))
+
+        moon_data = calcpos.Moon.calc(site, dt)
+        info.update(dict(
+            # Moon info here
+            moon_rise = (site.moon_rise(dt)).strftime("%H:%M:%S"),
+            moon_set = (site.moon_set(dt)).strftime("%H:%M:%S"),
+            moon_illum = str("%.2f%%" % ((site.moon_phase(
+                dt)) * 100)),
+            moon_ra = str(moon_data.ra),
+            moon_dec = str(moon_data.dec)))
+
+        return info
 
     def build_gui(self, container):
         obj = self.channel.opmon.get_plugin('SiteSelector')
-        self.site = obj.get_site().observer
-        self.dt_utc, self.tz = obj.get_datetime()
-
-        self.get_time_info()
-        self.get_sunmoon_info()
+        self.site_obj = obj.get_site()
+        self.dt_utc, self.cur_tz = obj.get_datetime()
+        obj.cb.add_callback('site-changed', self.site_changed_cb)
+        obj.cb.add_callback('time-changed', self.time_changed_cb)
 
         top = Widgets.VBox()
         top.set_border_width(4)
 
-        # Date info - TODO NEEDS LST
-        fr = Widgets.Frame('Date/Time Info')
+        # Date info - TODO: add LST
+        info = self.get_time_info()
+        fr = Widgets.Frame('Site Date/Time')
         self.w.dt_table = Widgets.GridBox(rows=3, columns=2)
         self.w.dt_table.add_widget(Widgets.Label('Date:'), 0, 0)
-        self.w.dt_table.add_widget(Widgets.Label('Local time:'), 1, 0)
-        self.w.dt_table.add_widget(Widgets.Label('UTC'), 2, 0)
-        self.w.dt_table.add_widget(Widgets.Label(self.date_current), 0, 1)
-        self.w.dt_table.add_widget(Widgets.Label(self.local_current), 1, 1)
-        self.w.dt_table.add_widget(Widgets.Label(self.utc), 2, 1)
+        self.w.dt_table.add_widget(Widgets.Label('Time:'), 1, 0)
+        self.w.dt_table.add_widget(Widgets.Label('UTC:'), 2, 0)
+        self.w.date_current = Widgets.Label(info.date_current)
+        self.w.dt_table.add_widget(self.w.date_current, 0, 1)
+        self.w.local_current = Widgets.Label(info.local_current)
+        self.w.dt_table.add_widget(self.w.local_current, 1, 1)
+        self.w.utc = Widgets.Label(info.utc)
+        self.w.dt_table.add_widget(self.w.utc, 2, 1)
 
         dt_hbox = Widgets.HBox()
         dt_hbox.add_widget(self.w.dt_table, stretch=0)
@@ -118,22 +124,30 @@ class PolarSky(GingaPlugin.LocalPlugin):
         fr.set_widget(dt_hbox)
         top.add_widget(fr, stretch=0)
 
+        info = self.get_sunmoon_info()
+
         # Sun info
         fr = Widgets.Frame('Sun')
-        self.w.sun_table = Widgets.GridBox(rows=4, columns=3)
-        self.w.sun_table.add_widget(Widgets.Label(''), 0, 0)
-        self.w.sun_table.add_widget(Widgets.Label('Site:'), 1, 0)
-        self.w.sun_table.add_widget(Widgets.Label('Nautical:'), 2, 0)
-        self.w.sun_table.add_widget(Widgets.Label('Astronomical:'), 3, 0)
-        self.w.sun_table.add_widget(Widgets.Label('Sunset'), 0, 1)
-        self.w.sun_table.add_widget(Widgets.Label(self.sun_set), 1, 1)
-        self.w.sun_table.add_widget(Widgets.Label(self.nautical_set), 2, 1)
-        self.w.sun_table.add_widget(Widgets.Label(self.astronomical_set), 3, 1)
-        self.w.sun_table.add_widget(Widgets.Label('Sunrise'), 0, 2)
-        self.w.sun_table.add_widget(Widgets.Label(self.sun_rise), 1, 2)
-        self.w.sun_table.add_widget(Widgets.Label(self.nautical_rise), 2, 2)
-        self.w.sun_table.add_widget(Widgets.Label(
-                                    self.astronomical_rise), 3, 2)
+        sun_table = Widgets.GridBox(rows=4, columns=3)
+        self.w.sun_table = sun_table
+        sun_table.add_widget(Widgets.Label(''), 0, 0)
+        sun_table.add_widget(Widgets.Label('Site:'), 1, 0)
+        sun_table.add_widget(Widgets.Label('Nautical:'), 2, 0)
+        sun_table.add_widget(Widgets.Label('Astronomical:'), 3, 0)
+        sun_table.add_widget(Widgets.Label('Sunset'), 0, 1)
+        self.w.sun_set = Widgets.Label(info.sun_set)
+        sun_table.add_widget(self.w.sun_set, 1, 1)
+        self.w.nautical_set = Widgets.Label(info.nautical_set)
+        sun_table.add_widget(self.w.nautical_set, 2, 1)
+        self.w.astronomical_set = Widgets.Label(info.astronomical_set)
+        sun_table.add_widget(self.w.astronomical_set, 3, 1)
+        sun_table.add_widget(Widgets.Label('Sunrise'), 0, 2)
+        self.w.sun_rise = Widgets.Label(info.sun_rise)
+        sun_table.add_widget(self.w.sun_rise, 1, 2)
+        self.w.nautical_rise = Widgets.Label(info.nautical_rise)
+        sun_table.add_widget(self.w.nautical_rise, 2, 2)
+        self.w.astronomical_rise = Widgets.Label(info.astronomical_rise)
+        sun_table.add_widget(self.w.astronomical_rise, 3, 2)
 
         sun_hbox = Widgets.HBox()
         sun_hbox.add_widget(self.w.sun_table, stretch=0)
@@ -143,17 +157,23 @@ class PolarSky(GingaPlugin.LocalPlugin):
 
         # Moon Info
         fr = Widgets.Frame('Moon')
-        self.w.moon_table = Widgets.GridBox(rows=2, columns=6)
-        self.w.moon_table.add_widget(Widgets.Label('Rise:'), 0, 0)
-        self.w.moon_table.add_widget(Widgets.Label('Set:'), 1, 0)
-        self.w.moon_table.add_widget(Widgets.Label('Illum:'), 2, 0)
-        self.w.moon_table.add_widget(Widgets.Label('RA:'), 3, 0)
-        self.w.moon_table.add_widget(Widgets.Label('DEC:'), 4, 0)
-        self.w.moon_table.add_widget(Widgets.Label(self.moon_rise), 0, 1)
-        self.w.moon_table.add_widget(Widgets.Label(self.moon_set), 1, 1)
-        self.w.moon_table.add_widget(Widgets.Label(self.moon_illum), 2, 1)
-        self.w.moon_table.add_widget(Widgets.Label(str(self.moon_RA)), 3, 1)
-        self.w.moon_table.add_widget(Widgets.Label(str(self.moon_DEC)), 4, 1)
+        moon_table = Widgets.GridBox(rows=2, columns=6)
+        self.w.moon_table = moon_table
+        moon_table.add_widget(Widgets.Label('Rise:'), 0, 0)
+        moon_table.add_widget(Widgets.Label('Set:'), 1, 0)
+        moon_table.add_widget(Widgets.Label('Illum:'), 2, 0)
+        moon_table.add_widget(Widgets.Label('RA:'), 3, 0)
+        moon_table.add_widget(Widgets.Label('DEC:'), 4, 0)
+        self.w.moon_rise = Widgets.Label(info.moon_rise)
+        moon_table.add_widget(self.w.moon_rise, 0, 1)
+        self.w.moon_set = Widgets.Label(info.moon_set)
+        moon_table.add_widget(self.w.moon_set, 1, 1)
+        self.w.moon_illum = Widgets.Label(info.moon_illum)
+        moon_table.add_widget(self.w.moon_illum, 2, 1)
+        self.w.moon_ra = Widgets.Label(info.moon_ra)
+        moon_table.add_widget(self.w.moon_ra, 3, 1)
+        self.w.moon_dec = Widgets.Label(info.moon_dec)
+        moon_table.add_widget(self.w.moon_dec, 4, 1)
 
         moon_hbox = Widgets.HBox()
         moon_hbox.add_widget(self.w.moon_table, stretch=0)
@@ -185,8 +205,6 @@ class PolarSky(GingaPlugin.LocalPlugin):
         return True
 
     def start(self):
-        self.update_table_timer_cb(self.tmr_table)
-        self.update_sunmoon_timer_cb(self.tmr_sunmoon)
         # self.viewer.set_bg(0.95, 0.95, 0.95)
         # self.viewer.set_fg(0.25, 0.25, 0.75)
 
@@ -223,7 +241,8 @@ class PolarSky(GingaPlugin.LocalPlugin):
         self.gui_up = False
         # remove the canvas from the image
         p_canvas = self.fitsimage.get_canvas()
-        p_canvas.delete_object(self.canvas)
+        if self.canvas in p_canvas:
+            p_canvas.delete_object(self.canvas)
 
     def redo(self):
         """This is called when a new image arrives or the data in the
@@ -231,16 +250,51 @@ class PolarSky(GingaPlugin.LocalPlugin):
         """
         pass
 
+    def site_changed_cb(self, cb, site_obj):
+        self.logger.debug("site has changed")
+        self.site_obj = site_obj
+        obj = self.channel.opmon.get_plugin('SiteSelector')
+        self.dt_utc, self.cur_tz = obj.get_datetime()
+        self.update_times()
+        self.update_sunmoon()
+
+    def time_changed_cb(self, cb, time_utc, cur_tz):
+        old_dt_utc = self.dt_utc
+        self.dt_utc = time_utc
+        self.cur_tz = cur_tz
+
+        elapsed = abs((self.dt_utc - old_dt_utc).total_seconds())
+        if elapsed > self.settings['times_update_interval']:
+            self.update_times()
+
+        if elapsed > 60 * 60 * 12:  # every 12 hours
+            self.update_sunmoon()
+
     def replot_all(self):
         self.initialize_plot()
 
-    def update_table_timer_cb(self, timer):
-        timer.start()
-        self.get_time_info()
+    def update_times(self):
+        info = self.get_time_info()
+        if self.gui_up:
+            self.w.date_current.set_text(info.date_current)
+            self.w.local_current.set_text(info.local_current)
+            self.w.utc.set_text(info.utc)
 
-    def update_sunmoon_timer_cb(self, timer):
-        timer.start()
-        self.get_sunmoon_info()
+    def update_sunmoon(self):
+        info = self.get_sunmoon_info()
+        if self.gui_up:
+            self.w.sun_set.set_text(info.sun_set)
+            self.w.nautical_set.set_text(info.nautical_set)
+            self.w.astronomical_set.set_text(info.astronomical_set)
+            self.w.sun_rise.set_text(info.sun_rise)
+            self.w.nautical_rise.set_text(info.nautical_rise)
+            self.w.astronomical_rise.set_text(info.astronomical_rise)
+
+            self.w.moon_rise.set_text(info.moon_rise)
+            self.w.moon_set.set_text(info.moon_set)
+            self.w.moon_illum.set_text(info.moon_illum)
+            self.w.moon_ra.set_text(info.moon_ra)
+            self.w.moon_dec.set_text(info.moon_dec)
 
     def change_radius_cb(self, setting, radius):
         self.replot_all()
@@ -343,7 +397,6 @@ class PolarSky(GingaPlugin.LocalPlugin):
         return scale
 
     def map_azalt(self, az, alt):
-        # az = subaru_normalize_az(az)
         return az + 90.0, 90.0 - alt
 
     def tel_posn_toggle_cb(self, w, tf):
